Update miniapps/services example
