+++ examples/miniapps/services/containers.py
@@ -2,6 +2,8 @@
 
 import sqlite3
 import boto.s3.connection
+
+import example.main
 import example.services
 
 import dependency_injector.containers as containers
@@ -31,3 +33,12 @@ class Services(containers.DeclarativeContainer):
     photos = providers.Factory(example.services.Photos,
                                db=Platform.database,
                                s3=Platform.s3)
+
+
+class Application(containers.DeclarativeContainer):
+    """IoC container of application component providers."""
+
+    main = providers.Callable(example.main.main,
+                              users_service=Services.users,
+                              auth_service=Services.auth,
+                              photos_service=Services.photos)
